+++ packages/atomic-angular/projects/atomic-angular/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.30.18](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.17...@coveo/atomic-angular@1.30.18) (2022-10-05)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.17](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.16...@coveo/atomic-angular@1.30.17) (2022-10-03)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.16](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.15...@coveo/atomic-angular@1.30.16) (2022-09-30)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.15](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.14...@coveo/atomic-angular@1.30.15) (2022-09-30)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.13...@coveo/atomic-angular@1.30.14) (2022-09-29)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.13](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.12...@coveo/atomic-angular@1.30.13) (2022-09-22)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.11...@coveo/atomic-angular@1.30.12) (2022-09-22)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
+## [1.30.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.10...@coveo/atomic-angular@1.30.11) (2022-09-21)
+
+**Note:** Version bump only for package @coveo/atomic-angular
+
+
+
+
+
 ## [1.30.10](https://github.com/coveo/ui-kit/compare/@coveo/atomic-angular@1.30.9...@coveo/atomic-angular@1.30.10) (2022-09-21)
 
 **Note:** Version bump only for package @coveo/atomic-angular