+++ packages/atomic-react/CHANGELOG.md
@@ -3,6 +3,70 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.24.19](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.18...@coveo/atomic-react@1.24.19) (2022-10-05)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.18](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.17...@coveo/atomic-react@1.24.18) (2022-10-03)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.17](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.16...@coveo/atomic-react@1.24.17) (2022-09-30)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.16](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.15...@coveo/atomic-react@1.24.16) (2022-09-30)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.15](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.14...@coveo/atomic-react@1.24.15) (2022-09-29)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.14](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.13...@coveo/atomic-react@1.24.14) (2022-09-22)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.13](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.12...@coveo/atomic-react@1.24.13) (2022-09-22)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
+## [1.24.12](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.11...@coveo/atomic-react@1.24.12) (2022-09-21)
+
+**Note:** Version bump only for package @coveo/atomic-react
+
+
+
+
+
 ## [1.24.11](https://github.com/coveo/ui-kit/compare/@coveo/atomic-react@1.24.10...@coveo/atomic-react@1.24.11) (2022-09-21)
 
 **Note:** Version bump only for package @coveo/atomic-react